requirements.txt
pydicom
numpy
matplotlib
scikit-image
pynrrd